requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.44
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.38.2
nltk==3.9.1
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pydeck==0.9.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
